requirements.txt
appdirs==1.4.3
asgiref==3.2.3
beautifulsoup4==4.8.2
celery==4.4.0
certifi==2019.11.28
cycler==0.10.0
Django==3.0.1
djangorestframework==3.11.0
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.8.1
django-filter==2.2.0
django-imagekit==4.0.2
django-mako-plus==5.11.2
django-polymorphic==2.1.2
django-sass-processor==0.7.5
django-select2==7.1.2
django-settings-export==1.2.1
django-taggit==1.2.0
django-taggit-labels==0.6.0
django-tinymce==2.8.0
django-typed-models==0.9.0
django-reverse-admin==2.8.0
kombu==4.6.7
mailchimp==2.0.9
Mako==1.1.0
markdown==3.1.1
natural-keys==1.5.1
pdfkit==0.6.1
pexpect==4.7.0
pipenv==2018.11.26
psycopg2-binary==2.8.4
ptyprocess==0.6.0
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.10.3
python-magic==0.4.15
pytz==2019.3
requests==2.22.0
selenium==3.141.0
social-auth-app-django==3.1.0
sqlparse==0.3.0
tabulate==0.8.6
termcolor==1.1.0
tinycss==0.4
virtualenv==16.7.9
virtualenv-clone==0.5.3
wcwidth==0.1.7
websocket-client==0.40.0